Возврат переплаты

+++ xbterminal/transactions/models.py
@@ -75,6 +75,10 @@ class Deposit(Transaction):
         max_digits=18,
         decimal_places=8,
         default=0)
+    refund_coin_amount = models.DecimalField(
+        max_digits=18,
+        decimal_places=8,
+        default=0)
 
     deposit_address = models.OneToOneField(
         'wallet.Address',
@@ -98,7 +102,6 @@ class Deposit(Transaction):
     time_broadcasted = models.DateTimeField(null=True)
     time_notified = models.DateTimeField(null=True)
     time_confirmed = models.DateTimeField(null=True)
-    time_refunded = models.DateTimeField(null=True)
     time_cancelled = models.DateTimeField(null=True)
 
     @property
@@ -132,7 +135,6 @@ class Deposit(Transaction):
                 reached the confidence threshold
             notified - customer notified about successful payment
             confirmed - incoming transaction confirmed
-            refunded - money sent back to customer
             timeout - incoming transaction did not received
             failed - incoming transaction received,
                 but payment order is not marked as notified
@@ -140,8 +142,6 @@ class Deposit(Transaction):
                 but incoming transaction is not confirmed
             cancelled: deposit cancelled by the customer
         """
-        if self.time_refunded:
-            return 'refunded'
         if self.time_cancelled:
             return 'cancelled'
         if self.time_notified:
@@ -185,16 +185,36 @@ class Deposit(Transaction):
 
     @atomic
     def create_balance_changes(self):
-        self.balancechange_set.all().delete()
-        self.balancechange_set.create(
+        if self.refund_coin_amount == self.paid_coin_amount:
+            # Full refund, delete balance changes
+            self.balancechange_set.all().delete()
+            return
+        self.balancechange_set.update_or_create(
             account=self.account,
             address=self.deposit_address,
-            amount=self.paid_coin_amount - self.fee_coin_amount)
+            amount__gt=0,
+            defaults={
+                'amount': self.paid_coin_amount - self.fee_coin_amount,
+            })
         if self.fee_coin_amount > 0:
-            self.balancechange_set.create(
-                account=None,
+            self.balancechange_set.update_or_create(
+                account__isnull=True,
                 address=self.deposit_address,
-                amount=self.fee_coin_amount)
+                amount__gt=0,
+                defaults={
+                    'amount': self.fee_coin_amount,
+                })
+        if self.refund_coin_amount > 0:
+            # Partial refund
+            if self.refund_coin_amount != self.paid_coin_amount - self.coin_amount:
+                raise ValueError
+            self.balancechange_set.update_or_create(
+                account=self.account,
+                address=self.deposit_address,
+                amount__lt=0,
+                defaults={
+                    'amount': -self.refund_coin_amount,
+                })
 
     @atomic
     def save(self, *args, **kwargs):
